Added microsoft code to Base to help maintencence. Refactored TestParamenterConfiguration to implement CommandLine rather than inherit from it.
