+++ src/Lucene.Net.TestFramework/Support/Configuration/ConfigurationBuilderExtensionMethods.cs
@@ -12,42 +12,18 @@ namespace Lucene.Net.Configuration
 {
     public static class ConfigurationBuilderExtensions
     {
+
         [CLSCompliant(false)]
-        public static List<IConfigurationSource> AddMicrosoftExtensionsConfigurationSource(this List<IConfigurationSource> sourceList , string TestBaseDirectory, string TestSettingsFileNameJson, string TestSettingsFileNameXml)
+        public static IConfigurationBuilder AddLuceneDefaultSettings(this IConfigurationBuilder configurationBuilder, string prefix)
         {
-            sourceList.Add(new TestParameterConfigurationSource(NUnit.Framework.TestContext.Parameters));
-
-            Stack<string> locations = ScanConfigurationFiles(TestBaseDirectory, TestSettingsFileNameJson);
-
-            while (locations.Count != 0)
-            {
-                sourceList.Add(new JsonConfigurationSource()
-                {
-                    Path = locations.Pop(),
-                    Optional = true,
-                    ReloadOnChange = true
-                });
-            }
-
-            locations = ScanConfigurationFiles(TestBaseDirectory, TestSettingsFileNameXml);
-
-            while (locations.Count != 0)
-            {
-                sourceList.Add(new XmlConfigurationSource()
-                {
-                    Path = locations.Pop(),
-                    Optional = true,
-                    ReloadOnChange = true
-                });
-            }
-            return sourceList;
+            configurationBuilder.Add(new LuceneDefaultConfigurationSource() { Prefix = prefix });
+            return configurationBuilder;
         }
-        
-
         [CLSCompliant(false)]
-        public static IConfigurationBuilder AddNUnitTestRunSettings(this IConfigurationBuilder builder)
+        public static IConfigurationBuilder AddNUnitTestRunSettings(this IConfigurationBuilder configurationBuilder)
         {
-            return builder.Add(new TestParameterConfigurationSource(NUnit.Framework.TestContext.Parameters));
+            configurationBuilder.Add(new TestParameterConfigurationSource() { TestParameters = NUnit.Framework.TestContext.Parameters });
+            return configurationBuilder;
         }
 
         [CLSCompliant(false)]
@@ -72,29 +48,6 @@ namespace Lucene.Net.Configuration
                 builder.AddXmlFile(locations.Pop(), optional: true, reloadOnChange: true);
             }
             return builder;
-        }
-
-
-        private static string FindParent(string currentPath, string fileName)
-        {
-            string candidatePath = System.IO.Path.Combine(currentPath, fileName);
-            try
-            {
-                while (new DirectoryInfo(currentPath).Parent != null)
-                {
-                    candidatePath = System.IO.Path.Combine(new DirectoryInfo(currentPath).Parent.FullName, fileName);
-                    if (File.Exists(candidatePath))
-                    {
-                        return candidatePath;
-                    }
-                    currentPath = new DirectoryInfo(currentPath).Parent.FullName;
-                }
-            }
-            catch (SecurityException)
-            {
-                // ignore security errors
-            }
-            return null;
         }
 
         private static Stack<string> ScanConfigurationFiles(string currentPath, string fileName)
