+++ src/Lucene.Net/Support/Configuration/Base/ConfigurationBuilder.cs
@@ -1,0 +1,71 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using Microsoft.Extensions.Configuration;
+using System;
+using System.Collections.Generic;
+
+namespace Lucene.Net.Configuration
+{
+    /// <summary>
+    /// Used to build key/value based configuration settings for use in an application.
+    /// </summary>
+    internal class ConfigurationBuilder : IConfigurationBuilder
+    {
+#if NETSTANDARD
+        /// <summary>
+        /// Returns the sources used to obtain configuration values.
+        /// </summary>
+        public IList<IConfigurationSource> Sources { get; } = new List<IConfigurationSource>();
+
+        /// <summary>
+        /// Gets a key/value collection that can be used to share data between the <see cref="IConfigurationBuilder"/>
+        /// and the registered <see cref="IConfigurationProvider"/>s.
+        /// </summary>
+        public IDictionary<string, object> Properties { get; } = new Dictionary<string, object>();
+#elif NET45
+        private Dictionary<string, object> _properties = new Dictionary<string, object>();
+
+        private IList<IConfigurationSource> _sources = new List<IConfigurationSource>();
+        public IEnumerable<IConfigurationSource> Sources { get { return _sources; } }
+
+        public Dictionary<string, object> Properties { get { return _properties; } }
+
+#endif
+        /// <summary>
+        /// Adds a new configuration source.
+        /// </summary>
+        /// <param name="source">The configuration source to add.</param>
+        /// <returns>The same <see cref="IConfigurationBuilder"/>.</returns>
+        public IConfigurationBuilder Add(IConfigurationSource source)
+        {
+            if (source == null)
+            {
+                throw new ArgumentNullException(nameof(source));
+            }
+
+#if NETSTANDARD
+            Sources.Add(source);
+#elif NET45
+            _sources.Add(source);
+#endif
+            return this;
+        }
+
+        /// <summary>
+        /// Builds an <see cref="IConfiguration"/> with keys and values from the set of providers registered in
+        /// <see cref="Sources"/>.
+        /// </summary>
+        /// <returns>An <see cref="IConfigurationRoot"/> with keys and values from the registered providers.</returns>
+        public IConfigurationRoot Build()
+        {
+            var providers = new List<IConfigurationProvider>();
+            foreach (var source in Sources)
+            {
+                var provider = source.Build(this);
+                providers.Add(provider);
+            }
+            return new ConfigurationRoot(providers);
+        }
+    }
+}
