+++ src/Lucene.Net/Support/Configuration/DefaultConfigurationFactory.cs
@@ -1,5 +1,7 @@
 ﻿using System;
+using System.Collections;
 using System.Collections.Generic;
+using System.Linq;
 using System.Security;
 using System.Text;
 using System.Threading;
@@ -8,7 +10,7 @@ using Microsoft.Extensions.Primitives;
 
 namespace Lucene.Net.Configuration
 {
-    public class DefaultConfigurationFactory : IConfigurationFactory
+    internal class DefaultConfigurationFactory : IConfigurationFactory
     {
         private readonly bool ignoreSecurityExceptionsOnRead;
         private bool initialized = false;
@@ -18,12 +20,11 @@ namespace Lucene.Net.Configuration
 
         public DefaultConfigurationFactory(bool ignoreSecurityExceptionsOnRead)
         {
-            this.builder = new LuceneConfigurationBuilder();
-            builder.Add(new LuceneConfigurationSource() { Prefix = "lucene:" });
+            this.builder = new ConfigurationBuilder();
+            builder.Add(new LuceneDefaultConfigurationSource() { Prefix = "lucene:" });
             this.ignoreSecurityExceptionsOnRead = ignoreSecurityExceptionsOnRead;
         }
 
-        [CLSCompliant(false)]
         public virtual IConfiguration CreateConfiguration()
         {
             return EnsureInitialized();
@@ -33,7 +34,6 @@ namespace Lucene.Net.Configuration
         /// Ensures the <see cref="Initialize"/> method has been called since the
         /// last application start. This method is thread-safe.
         /// </summary>
-        [CLSCompliant(false)]
         protected IConfiguration EnsureInitialized()
         {
             return LazyInitializer.EnsureInitialized(ref this.configuration, ref this.initialized, ref this.m_initializationLock, () =>
@@ -46,13 +46,11 @@ namespace Lucene.Net.Configuration
         /// <summary>
         /// Initializes the dependencies of this factory.
         /// </summary>
-        [CLSCompliant(false)]
         protected virtual IConfiguration Initialize()
         {
             return builder.Build();
         }
     }
-
     //internal class DefaultConfiguration : IConfiguration
     //{
     //    private readonly bool ignoreSecurityExceptionsOnRead;
